requirements.txt
# Библиотека для работы с изображениями
Pillow==10.1.0

# Загрузка переменных окружения из файла .env
python-dotenv==1.0.0

# Библиотека для HTTP запросов
requests==2.31.0

# Библиотека для работы с OpenAI API
openai>=1.0.0

# Библиотека для работы с GUI
tk==0.1.0